fixed the dll warnings and some other warning (#56)
warnings that are coming from dependency some fixed some ignored
DLL, only to export the public functions that is gonna use in editor side
you can export out the entire class as long as there is no  use of std:: classes

+++ GAM300Engine/Include/ecs/ecs.hpp
@@ -431,7 +431,7 @@ namespace TDS
         // Add entity ID to new archetype
         newArchetype->entityIds.emplace_back(entityID);
         // Change index
-        record.index = newArchetype->entityIds.size() - 1;
+        record.index = static_cast<uint32_t>(newArchetype->entityIds.size() - 1);
         // Change archetype pointer
         record.archetype = newArchetype;
 
@@ -582,7 +582,7 @@ namespace TDS
         oldArchetype->entityIds.erase(willBeRemoved);
 
         newArchetype->entityIds.emplace_back(entityID);
-        record.index = newArchetype->entityIds.size() - 1;
+        record.index = static_cast<uint32_t>(newArchetype->entityIds.size() - 1);
         record.archetype = newArchetype;
     }
